Move member.h to syntax/ and wrap it inside syntax namespace
