+++ include/absent/combinators/fmap.h
@@ -1,7 +1,7 @@
 #ifndef RVARAGO_ABSENT_FMAP_H
 #define RVARAGO_ABSENT_FMAP_H
 
-#include "absent/member.h"
+#include "absent/syntax/member.h"
 #include "absent/syntax/nullable.h"
 
 #include <functional>
@@ -31,7 +31,7 @@ namespace rvarago::absent {
      * The same as fmap but for a member function that has to be const and parameterless.
      */
     template <typename A, typename B, template <typename> typename Nullable>
-    constexpr auto fmap(Nullable<A> const& input, member::Mapper<const A, B> fn) -> Nullable<B> {
+    constexpr auto fmap(Nullable<A> const& input, syntax::member::Mapper<const A, B> fn) -> Nullable<B> {
         return fmap(input, [&fn](auto const& input_value){ return std::invoke(fn, input_value); });
     }
 
@@ -47,7 +47,7 @@ namespace rvarago::absent {
      * Infix version of fmap for a member function.
      */
     template <typename A, typename B, template <typename> typename Nullable>
-    constexpr auto operator|(Nullable<A> const& input, member::Mapper<const A, B> fn) -> Nullable<B> {
+    constexpr auto operator|(Nullable<A> const& input, syntax::member::Mapper<const A, B> fn) -> Nullable<B> {
         return fmap(input, fn);
     }
